scripts/requirements.txt
kagglehub
fredapi
pandas
numpy
python-dotenv
